football_env/pyproject.toml
[project]
name = "football-env"
version = "0.1.0"
description = "NFL play-calling RL environment for OpenEnv"
requires-python = ">=3.10"
dependencies = ["openenv-core>=0.2.1", "uvicorn>=0.30.0"]

[project.scripts]
server = "football_env.server.app:main"
